+++ SpringWebApp/src/main/java/isamrs/service/LekarServiceImpl.java
@@ -27,7 +27,7 @@ public class LekarServiceImpl implements LekarService {
 
 	@Override
 	public Lekar create(Lekar lekar) {
-		Long id = (long) this.findAll().size();
+		Integer id = (int) this.findAll().size();
 		lekar.setId(id);
 		return lekarRepo.create(lekar);
 	}
